unl0kr: allow runtime selection of the lvgl display driver ("backend")
lvgl supports multiple display drivers, and it's safe to ship more than one
in the final executable [0].

This commit adds plumbing to support runtime selection of the lvgl display
driver. Apart from the already supported fbdev driver, this commit also
adds support for the DRM driver.

The driver (herein referred as "backend") can be modified via the
general.backend configuration.

The eventual DRM backend can be enabled in `lv_drv_conf.h`,
otherwise it won't be built and it won't be available in unl0kr.

A subsequent commit will expose the feature as a meson build option.

[0] https://docs.lvgl.io/latest/en/html/overview/display.html

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>