HACK: indev: raise MAX_TOUCHSCREEN_DEVS to 2
Some devices (such as Sony Xperia 5) ship another touchscreen on the
device's sides ("side sense"). This isn't used on unl0kr, but it might
be picked up by mistake, skipping the detection of the real touchscreen.

This commit raises the touchscreen ceiling to 2 so that they both can be
deteted.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>